Subindo notificações e avaliações finalizadas

+++ app/src/main/java/com/stack/open_work_mobile/services/AvaliationService.kt
@@ -9,14 +9,14 @@ import retrofit2.http.Path
 interface AvaliationService {
 
     @GET("/api/avaliacoes/desenvolvedor/{id}")
-    fun getAvaliationsDev(@Path("id") id: Int): Call<RatingCompanies>
+    fun getAvaliationsDev(@Path("id") id: Long): Call<RatingCompanies>
 
     @GET("/api/avaliacoes/empresa/{id}")
-    fun getAvaliationsCompany(@Path("id") id: Int): Call<RatingCompanies>
+    fun getAvaliationsCompany(@Path("id") id: Long): Call<RatingCompanies>
 
     @POST("/api/empresa/{id}/{grade}")
     fun registerAvaliationDeveloper(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Path("grade") grade: Int
     ): Call<RatingCompanies>
